+++ ui/packages/app/grafana-datasource-plugin/src/datasource.ts
@@ -30,6 +30,7 @@ import {
   QueryServiceClient,
   HealthClient,
   HealthCheckResponse_ServingStatus,
+  Label,
 } from '@parca/client';
 import { saveAsBlob } from '@parca/functions';
 
@@ -68,7 +69,7 @@ export class DataSource extends DataSourceApi<ParcaQuery, ParcaDataSourceOptions
           refId: query.refId,
           fields: [{ name: 'data', type: FieldType.other }],
         });
-        frame.appendRow([await this.getData(from, to, query)]);
+        frame.appendRow([await this.getData(from, to, query, [])]);
         return frame;
       })
     );
@@ -76,8 +77,8 @@ export class DataSource extends DataSourceApi<ParcaQuery, ParcaDataSourceOptions
     return { data };
   }
 
-  async getData(from: number, to: number, query: ParcaQuery): Promise<GrafanaParcaData> {
-    const profileSource = new MergedProfileSource(from, to, query.parcaQuery);
+  async getData(from: number, to: number, query: ParcaQuery, labels: Label[]): Promise<GrafanaParcaData> {
+    const profileSource = new MergedProfileSource(from, to, labels, query.parcaQuery);
     const flamegraphReq = profileSource.QueryRequest();
     flamegraphReq.reportType = QueryRequest_ReportType.FLAMEGRAPH_TABLE;
     const topTableReq = profileSource.QueryRequest();