+++ ui/packages/app/grafana-panel-plugin/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.68 (2023-02-15)
+
+**Note:** Version bump only for package @parca/grafana-panel
+
 ## [0.0.67](https://github.com/parca-dev/parca/compare/@parca/grafana-panel@0.0.66...@parca/grafana-panel@0.0.67) (2023-02-13)
 
 **Note:** Version bump only for package @parca/grafana-panel