+++ ui/packages/app/web/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.202](https://github.com/parca-dev/parca/compare/@parca/web@0.16.201...@parca/web@0.16.202) (2023-02-15)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.201](https://github.com/parca-dev/parca/compare/@parca/web@0.16.200...@parca/web@0.16.201) (2023-02-15)
+
+**Note:** Version bump only for package @parca/web
+
 ## [0.16.200](https://github.com/parca-dev/parca/compare/@parca/web@0.16.199...@parca/web@0.16.200) (2023-02-14)
 
 **Note:** Version bump only for package @parca/web